src/front-end/src/python/requirements.txt
pytesseract
colorthief
opencv-python
Pillow
onnxruntime
fast-plate-ocr
open-image-models
fastapi
uvicorn
python-multipart
scikit-learn
numpy

